+++ include/mq/imgui/Widgets.h
@@ -28,14 +28,16 @@ namespace mq::imgui {
 // Draws a CTextureAnimation. This is the preferred method for drawing a UI texture to imgui.
 // This function uses the provided size, if available. Otherwise, it uses the internally
 // specified size of the texture.
-MQLIB_OBJECT bool DrawTextureAnimation(const eqlib::CTextureAnimation* textureAnimation, const eqlib::CXSize& size = eqlib::CXSize());
+MQLIB_OBJECT bool DrawTextureAnimation(const eqlib::CTextureAnimation* textureAnimation, const eqlib::CXSize& size = eqlib::CXSize(),
+	bool drawBorder = false);
 
 // Draws a CUITextureInfo. This is the lowest level of the UI texture hierarchy, and represents a full individual texture.
-MQLIB_OBJECT bool DrawUITexture(const eqlib::CUITextureInfo& textureInfo, const eqlib::CXRect& rect = CXRect(0, 0, -1, -1), const eqlib::CXSize& size = CXSize());
+MQLIB_OBJECT bool DrawUITexture(const eqlib::CUITextureInfo& textureInfo, const eqlib::CXRect& rect = CXRect(0, 0, -1, -1),
+	const eqlib::CXSize& size = CXSize(), bool drawBorder = false);
 
 // Draws a TexturePiece. srcRect can be used to draw only a sub-section of the full texture.
-MQLIB_OBJECT bool DrawTexturePiece(const eqlib::CUITexturePiece& texturePiece, const CXRect& srcRect, const CXSize& imageSize);
-MQLIB_OBJECT bool DrawTexturePiece(const eqlib::CUITexturePiece& texturePiece, const CXSize& imageSize);
+MQLIB_OBJECT bool DrawTexturePiece(const eqlib::CUITexturePiece& texturePiece, const CXRect& srcRect, const CXSize& imageSize, bool drawBorder = false);
+MQLIB_OBJECT bool DrawTexturePiece(const eqlib::CUITexturePiece& texturePiece, const CXSize& imageSize, bool drawBorder = false);
 
 // Draws a text link styled like an item link. Returns true if the link has been clicked.
 MQLIB_API bool ItemLinkText(const char* fmt, ...);